Move GListItemAdapter into 'adapters' directory
